LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/ready-states/autoplay-with-slow-text-tracks-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'track.track.cues.length')
CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'track.track.cues.length')
CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'track.track.cues.length')
CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'track.track.cues.length')
CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'track.track.cues.length')

FAIL autoplay with slow text tracks TypeError: null is not an object (evaluating 'track.track.cues.length')
